requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.32
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
openai==1.50.0
httpx==0.27.2
pydantic==2.9.0
bcrypt==4.0.1
python-multipart==0.0.12
psycopg2-binary==2.9.9
firebase-admin==6.5.0
